+++ controllers/user.js
@@ -6,6 +6,7 @@
 'use strict';
 
 var fs = require('fs');
+var fetchZipcode = require('../lib/fetch_zipcode');
 var cloudinary = require('cloudinary');
 
 var User = require('../models/user'); // User Model
@@ -14,7 +15,10 @@ var INFO = [
   'User deleted',
   'User information updated',
   'User location updated',
-  'User profile updated'
+  'User profile updated',
+  'New blocked user added',
+  'User already blocked',
+  'Removed blocked users'
 ];
 
 var ERROR = [
@@ -23,12 +27,14 @@ var ERROR = [
   'User validation failed'
 ];
 
-// GET /api/users/:username
-exports.getUserInfo = function (req, res, next) {
-  User.findOne({ username: req.params.username }, {
-    password: 0,
-    __v: 0,
-    updatedAt: 0
+var USER_ID_PATTERN = /^[0-9a-fA-F]{24}$/;
+
+/* GET /api/users/:username */
+exports.getUser = function (req, res, next) {
+  User.findOne({ 'username': req.params.username }, {
+    'password': 0,
+    '__v': 0,
+    'updatedAt': 0
   }).lean().exec(function (err, user) {
 
     if (err) { return next(err); }
@@ -49,10 +55,12 @@ exports.getUserInfo = function (req, res, next) {
   });
 };
 
-// DELETE /api/users/:userid
+/* DELETE /api/users/:user_id */
 exports.deleteUser = function (req, res, next) {
-  if (req.params.userid.match(/^[0-9a-fA-F]{24}$/)) {
-    User.remove({ _id: req.params.userid }, function (err, user) {
+  if (!req.params.user_id.match(USER_ID_PATTERN)) {
+    res.status(400).json({ message: ERROR[1] });
+  } else {
+    User.remove({ '_id': req.params.user_id }, function (err, user) {
 
       if (err) { return next(err); }
 
@@ -66,20 +74,19 @@ exports.deleteUser = function (req, res, next) {
         });
       }
     });
-  } else {
-    res.status(400).json({
-      message: ERROR[1]
-    });
   }
 };
 
-// PUT /api/users/:userid
-exports.updateUserInfo = function (req, res, next) {
-  if (req.params.userid.match(/^[0-9a-fA-F]{24}$/)) {
+/* PUT /api/users/:user_id */
+exports.updateUser = function (req, res, next) {
+  if (!req.params.user_id.match(USER_ID_PATTERN)) {
+    res.status(400).json({
+      message: ERROR[1]
+    });
+  } else {
+    var NEW_USER_INFO = req.body; // the received user object
 
-    var userInfo = req.body;
-
-    User.findOne({ _id: req.params.userid }, function (err, user) {
+    User.findOne({ '_id': req.params.user_id }, function (err, user) {
 
       if (err) { return next(err); }
 
@@ -89,9 +96,10 @@ exports.updateUserInfo = function (req, res, next) {
         });
       } else if (user) {
 
-        user.username = userInfo.username || user.username;
-        user.password = userInfo.password || user.password;
-        user.email = userInfo.email || user.email;
+        user.active = NEW_USER_INFO.active? true : false;
+        user.username = NEW_USER_INFO.username || user.username;
+        user.password = NEW_USER_INFO.password || user.password;
+        user.email = NEW_USER_INFO.email || user.email;
 
         // catch validation error and return response to the client
         user.save(function (err) {
@@ -107,19 +115,18 @@ exports.updateUserInfo = function (req, res, next) {
         });
       }
     });
-  } else {
-    res.status(400).json({
-      message: ERROR[1]
-    });
   }
 };
 
-// POST /api/users/:userid/location
+/* POST /api/users/:user_id/location */
 exports.updateUserLocation = function (req, res, next) {
-  if (req.params.userid.match(/^[0-9a-fA-F]{24}$/)) {
+  var coords = [req.body.lng, req.body.lat];
+  var zipcode = fetchZipcode(coords);
+
+  if (req.params.user_id.match(/^[0-9a-fA-F]{24}$/)) {
 
     // store new coordenates
-    User.findOne({ _id: req.params.userid }, function (err, user) {
+    User.findOne({ '_id': req.params.user_id }, function (err, user) {
 
       if (err) { return next(err); }
 
@@ -129,7 +136,8 @@ exports.updateUserLocation = function (req, res, next) {
         });
       } else if (user) {
 
-        user.loc = [req.body.lng, req.body.lat];
+        user.loc = coords; // set new coordinates
+        user.loc_attr.zipcode = zipcode; // update zipcode
 
         user.save(function (err) {
           if (err) {
@@ -151,15 +159,20 @@ exports.updateUserLocation = function (req, res, next) {
   }
 };
 
-// PUT /api/users/:userid/location
 // NOTE this method returns an empty array if no users were found with nearby coordinates
+/* PUT /api/users/:user_id/location */
 exports.getNearbyUsers = function (req, res, next) {
-  if (req.params.userid.match(/^[0-9a-fA-F]{24}$/)) {
+  if (!req.params.user_id.match(USER_ID_PATTERN)) {
+    res.status(400).json({
+      message: ERROR[1]
+    });
+  } else {
 
-    var userid = req.params.userid;
+    var coords = [req.body.lng, req.body.lat];
+    var zipcode = fetchZipcode(coords);
 
     // store new coordinates
-    User.findOne({ _id: userid }, function (err, user) {
+    User.findOne({ '_id': req.params.user_id }, function (err, user) {
 
       if (err) { return next(err); }
 
@@ -169,7 +182,8 @@ exports.getNearbyUsers = function (req, res, next) {
         });
       } else if (user) {
 
-        user.loc = [req.body.lng, req.body.lat]; // add new coordinates
+        user.loc = coords;
+        user.loc_attr.zipcode = zipcode;
 
         // save changes
         user.save(function (err) {
@@ -186,33 +200,52 @@ exports.getNearbyUsers = function (req, res, next) {
         // convert the distance to radius
         maxDistance /= 6371;
 
-        var coords = [];
-        coords[0] = req.body.lng;
-        coords[1] = req.body.lat;
+        // inactivity max time limit value (in hours);
+        var inactiveTimeLimit = 5;
 
         // query for nearby users
-        User.find({
-          loc: {
-            $near: coords,
-            $maxDistance: maxDistance
+        User.aggregate([
+          {
+            '$geoNear': {
+              'near': coords,
+              'distanceField': 'calculated_distance',
+              'maxDistance': maxDistance,
+              'spherical': false
+            }
+          },
+          {
+            '$match': {
+              '_id': {
+                '$nin': user.blocked_users
+              },
+              'updatedAt': {
+                '$gte': new Date(new Date().setHours(new Date().getHours() - inactiveTimeLimit)),
+                '$lte': new Date()
+              },
+              'loc_attr.zipcode': zipcode
+            }
+          },
+          {
+            '$project': {
+              '_id': 1,
+              'updatedAt': 1,
+              'active': 1,
+              'username': 1,
+              'loc': 1,
+              'profile': 1
+            }
           }
-        }, {
-          password: 0,
-          __v: 0,
-          updatedAt: 0
-        }).lean().exec(function (err, users) {
+        ], function (err, users) {
           if (err) { return next(err); }
 
           // this removes the current user from the results array
-          // NOTE mongoose doesn't provide this type of functionality (excluding a specific user from a query)
-          // looping through the users array and eliminating the user's object is the only feasible solution I could find
           users.forEach(function (elem, index) {
-            if (elem._id == userid) {
+            if (elem._id == req.params.user_id) {
               users.splice(index, 1);
             }
           });
 
-          // NOTE override location property
+          // modifies location properties and was_active string
           for (var key in users) {
             users[key].loc = {
               lng: users[key].loc[0],
@@ -224,44 +257,16 @@ exports.getNearbyUsers = function (req, res, next) {
         });
       }
     });
-  } else {
-    res.status(400).json({
-      message: ERROR[1]
-    });
   }
 };
 
-// GET /api/users/:username/profile
-exports.getUserProfile = function (req, res, next) {
-  User.findOne({ username: req.params.username }, {
-    // remove unwanted or sensitive fields
-    password: 0,
-    email: 0,
-    __v: 0,
-    loc: 0,
-    updatedAt: 0,
-    createdAt: 0
-  }, function (err, profile) {
-
-    if (err) { return next(err); }
-
-    if (!profile) {
-      res.status(404).json({
-        message: ERROR[0]
-      });
-    } else {
-      res.status(200).json(profile);
-    }
-  });
-};
-
-// PUT /api/users/:username/profile
 // NOTE upload image to cloudinary from the client
+/* PUT /api/users/:username/profile */
 exports.updateUserProfile = function (req, res, next) {
 
-  var newProfileInfo = req.body;
+  var NEW_PROFILE_INFO = req.body;
 
-  User.findOne({ username: req.params.username }, function (err, user) {
+  User.findOne({ 'username': req.params.username }, function (err, user) {
 
     if (err) { return next(err); }
 
@@ -289,8 +294,8 @@ exports.updateUserProfile = function (req, res, next) {
         });
       }
 
-      user.profile.gender = newProfileInfo.gender || user.profile.gender;
-      user.profile.bio = newProfileInfo.bio || user.profile.bio;
+      user.profile.gender = NEW_PROFILE_INFO.gender || user.profile.gender;
+      user.profile.status = NEW_PROFILE_INFO.status || user.profile.status;
 
       user.save(function (err) {
         if (err) { return next(err); }
@@ -303,3 +308,77 @@ exports.updateUserProfile = function (req, res, next) {
   });
 };
 
+
+exports.blockUser = function (req, res, next) {
+  if (!req.params.user_id.match(USER_ID_PATTERN)) {
+    res.status(400).json({
+      message: ERROR[1]
+    });
+  } else {
+
+    var BLOCKED_USER_ID = req.query.blocked_user;
+
+    User.findOne({ '_id': req.params.user_id }, function (err, user) {
+      if (err) { return next(err); }
+
+      if (!user) {
+        res.status(404).send({
+          message: ERROR[0]
+        });
+      } else if (user) {
+
+        if (user.blocked_users.indexOf(BLOCKED_USER_ID) > -1) {
+          res.status(200).json({
+            message: INFO[5]
+          });
+        } else {
+          user.blocked_users.push(BLOCKED_USER_ID);
+
+          user.save(function (err) {
+            if (err) { return next(err); }
+          });
+
+          res.status(200).json({
+            message: INFO[4]
+          });
+        }
+      }
+    });
+  }
+};
+
+exports.removeBlockedUsers = function (req, res, next) {
+  if (!req.params.user_id.match(USER_ID_PATTERN)) {
+    res.status(400).json({
+      message: ERROR[1]
+    });
+  } else {
+    var BLOCKED_USERS = JSON.parse(req.query.blocked_users);
+
+    User.findOne({ '_id': req.params.user_id }, function (err, user) {
+      if (err) { return next(err); }
+
+      if (!user) {
+        res.status(404).send({
+          message: ERROR[0]
+        });
+      } else if (user) {
+        user.blocked_users.forEach(function (addedBlockedUsers) {
+          BLOCKED_USERS.forEach(function (toRemove) {
+            if (addedBlockedUsers == toRemove) {
+              user.blocked_users.splice(addedBlockedUsers, 1);
+            }
+          });
+        });
+
+        user.save(function (err) {
+          if (err) { return next(err); }
+        });
+
+        res.status(200).json({
+          message: INFO[6]
+        });
+      }
+    });
+  }
+};