+++ test/test.js
@@ -6,11 +6,11 @@ var should = require('should');
 var request = require('supertest');
 
 var testUser = {
-  _id: '56e3001a76517d7d05122b1b',
+  _id: '5705b27cfe67e893100c262d',
   username: 'user1',
   password: '1234',
   email: 'user1@server.com',
-  loc: [-69.758605, 19.475206],
+  loc: [-66.75861, 18.47524],
   profile: {
     profile_image: undefined,
     gender: 'Male',
@@ -43,7 +43,10 @@ var server_test = function (server) {
           username: testUser.username,
           password: testUser.password,
           email: testUser.email,
-          "loc": [testUser.loc[0], testUser.loc[1]]
+          loc: {
+            lng: testUser.loc[0],
+            lat: testUser.loc[1]
+          }
       };
 
       request(url)
@@ -334,21 +337,6 @@ var server_test = function (server) {
           .set('Accept', 'application/json')
           .expect('Content-Type', /json/)
           .send(payload)
-          .end(function (err, res) {
-            try {
-              if (err) { throw err; }
-
-              should(res).have.property('status', 200);
-              done();
-            } catch (e) {
-              done(e);
-            }
-          });
-      });
-
-      it('should RETURN a user\'s profile information', function (done) {
-        request(url)
-          .get('/api/users/' + testUser.username + '/profile')
           .end(function (err, res) {
             try {
               if (err) { throw err; }