+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 1.3.1
+
+- Adds nearby users query filters (by zipcode and last_updated properties)
+- Minor fixes and improvements
+
+# 1.3.0
+
+- Adds socket.io support
+- Changes geospatial query operator from `$near` to `$geoNear` for MongoDB replica sets and sharding support
+
 # 1.2.0
 
 - Created authentication_tests branch